# onyx-procurement/docs/openapi.yaml
openapi: 3.0.3
info:
  title: ONYX Procurement API
  description: |
    API for the ONYX Procurement System — covers procurement (suppliers,
    RFQ, quotes, POs, subcontractors), VAT (PCN836), annual tax forms,
    bank reconciliation, payroll/wage slips, and WhatsApp webhooks.
  version: 1.0.0
  contact:
    name: Techno-Kol Ops
servers:
  - url: http://localhost:3100
    description: Local development
  - url: https://api.onyx.example.com
    description: Production placeholder
...
security:
  - ApiKeyAuth: []
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
...
    Project:
      type: object
      additionalProperties: true
    Customer:
      type: object
      additionalProperties: true
    CustomerInvoice:
      type: object
      additionalProperties: true
    CustomerPayment:
      type: object
      additionalProperties: true
      properties:
        id: { type: integer }
        customer_id: { type: integer }
        amount: { type: number }
        payment_date: { type: string, format: date }
        receipt_number: { type: string }
        invoice_ids:
          type: array
          items: { type: integer }
    FiscalYear:
      type: object
      additionalProperties: true
    AnnualTaxReport:
      type: object
      additionalProperties: true
...
  responses:
...
    BadRequest:
      description: Invalid request payload or parameters
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
paths:
...
  /api/projects:
    get:
      tags: [Annual Tax]
      summary: List projects
      parameters:
        - in: query
          name: status
          schema: { type: string }
        - in: query
          name: fiscal_year
          schema: { type: integer }
      responses:
        "200":
          description: List
          content:
            application/json:
              schema:
                type: object
                properties:
                  projects:
                    type: array
                    items: { $ref: "#/components/schemas/Project" }
    post:
      tags: [Annual Tax]
      summary: Create a project
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/Project" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                type: object
                properties:
                  project: { $ref: "#/components/schemas/Project" }
  /api/projects/{id}:
    parameters:
      - in: path
        name: id
        required: true
        schema: { type: string }
    patch:
      tags: [Annual Tax]
      summary: Update a project
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/Project" }
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  project: { $ref: "#/components/schemas/Project" }
  /api/customers:
    get:
      tags: [Annual Tax]
      summary: List active customers
      responses:
        "200":
          description: List
          content:
            application/json:
              schema:
                type: object
                properties:
                  customers:
                    type: array
                    items: { $ref: "#/components/schemas/Customer" }
    post:
      tags: [Annual Tax]
      summary: Create customer
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/Customer" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                type: object
                properties:
                  customer: { $ref: "#/components/schemas/Customer" }
  /api/customer-invoices:
    get:
      tags: [Annual Tax]
      summary: List customer invoices
      parameters:
        - in: query
          name: customer_id
          schema: { type: string }
        - in: query
          name: project_id
          schema: { type: string }
        - in: query
          name: status
          schema: { type: string }
        - in: query
          name: limit
          schema: { type: integer, default: 100 }
      responses:
        "200":
          description: List
          content:
            application/json:
              schema:
                type: object
                properties:
                  invoices:
                    type: array
                    items: { $ref: "#/components/schemas/CustomerInvoice" }
    post:
      tags: [Annual Tax]
      summary: Create a customer invoice (auto VAT)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CustomerInvoice" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                type: object
                properties:
                  invoice: { $ref: "#/components/schemas/CustomerInvoice" }
  /api/customer-payments:
    get:
      tags: [Annual Tax]
      summary: List customer payments
      parameters:
        - in: query
          name: limit
          schema: { type: integer, default: 100 }
      responses:
        "200":
          description: List
          content:
            application/json:
              schema:
                type: object
                properties:
                  payments:
                    type: array
                    items: { $ref: "#/components/schemas/CustomerPayment" }
    post:
      tags: [Annual Tax]
      summary: Record a customer payment (optionally apply to invoices)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CustomerPayment" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                type: object
                properties:
                  payment: { $ref: "#/components/schemas/CustomerPayment" }
...
  /api/fiscal-years/{year}/compute:
    parameters:
      - in: path
        name: year
        required: true
        schema: { type: integer }
    post:
      tags: [Annual Tax]
      summary: Compute revenue/COGS/profit for the year
      responses:
        "200":
          description: Fiscal year upserted
          content:
            application/json:
              schema:
                type: object
                properties:
                  fiscal_year: { $ref: "#/components/schemas/FiscalYear" }
        "400": { $ref: "#/components/responses/BadRequest" }
  /api/annual-tax/{year}/forms/{type}/generate:
    parameters:
      - in: path
        name: year
        required: true
        schema: { type: integer }
      - in: path
        name: type
        required: true
        schema: { type: string, enum: ["1320", "1301", "6111", "30a"] }
    post:
      tags: [Annual Tax]
      summary: Generate an annual tax form
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        "200":
          description: Report generated
          content:
            application/json:
              schema:
                type: object
                properties:
                  report: { $ref: "#/components/schemas/AnnualTaxReport" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "412":
          description: Profile or fiscal year missing
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /api/annual-tax/{year}/forms:
    parameters:
      - in: path
        name: year
        required: true
        schema: { type: integer }
    get:
      tags: [Annual Tax]
      summary: List annual tax reports for the year
      responses:
        "200":
          description: List
          content:
            application/json:
              schema:
                type: object
                properties:
                  reports:
                    type: array
                    items: { $ref: "#/components/schemas/AnnualTaxReport" }
...